# docs/swagger.yaml
swagger: "2.0"
basePath: '{{.BasePath}}'
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
...
paths:
...
  /v1/plaid/create_link_token:
    get:
      tags:
        - plaid
      description: Plaid Create Link Token 
      summary: Plaid Link token
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: header
          name: Authorization
          required: true
          type: string
          format: byte
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/LinkTokenResponse'
        500:
          description: Client error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /v1/plaid/set_access_token:
    post:
      tags:
        - plaid
      description: Plaid Set Public Token and Account ID
      summary:  Plaid Set Public Token and Account ID
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: public_token
          in: body
          required: true
          description: Public token
          schema:
            type: string
        - name: account_id
          in: body
          required: true
          description: Selected Account Bank ID
          schema:
            type: string
        - in: header
          name: Authorization
          required: true
          type: string
          format: byte
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/AccessTokenResponse'
        500:
          description: Client error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
      errors:
        type: object
        items:
          type: object
...
  AccessTokenResponse:
     type: object
     properties:
       id:
         type: integer
       public_token:
         type: string
       item_id:
         type: string
...
  LinkTokenResponse:
    type: object
    properties:
      link_token:
        type: string
...